Revert "Rename djangosaml2 to djangosaml2_v0181post"
