+++ djangosaml2/conf.py
@@ -21,7 +21,7 @@ from django.core.exceptions import ImproperlyConfigured
 
 from saml2.config import SPConfig
 
-from djangosaml2_v0181post.utils import get_custom_setting
+from djangosaml2.utils import get_custom_setting
 
 
 def get_config_loader(path, request=None):
@@ -64,7 +64,7 @@ def config_settings_loader(request=None):
 
 def get_config(config_loader_path=None, request=None):
     config_loader_path = config_loader_path or get_custom_setting(
-        'SAML_CONFIG_LOADER', 'djangosaml2_v0181post.conf.config_settings_loader')
+        'SAML_CONFIG_LOADER', 'djangosaml2.conf.config_settings_loader')
 
     config_loader = get_config_loader(config_loader_path)
     return config_loader(request)
