+++ djangosaml2/exceptions.py
