+++ djangosaml2/signals.py
