+++ djangosaml2/tests/attribute-maps/saml_uri.py
