+++ djangosaml2/tests/remote_metadata_three_idps.xml
