+++ djangosaml2/utils.py
@@ -43,7 +43,7 @@ def get_idp_sso_supported_bindings(idp_entity_id=None, config=None):
     This is not clear in the pysaml2 code, so wrapping it in a util"""
     if config is None:
         # avoid circular import
-        from djangosaml2_v0181post.conf import get_config
+        from djangosaml2.conf import get_config
         config = get_config()
     # load metadata store from config
     meta = getattr(config, 'metadata', {})
@@ -78,7 +78,7 @@ def fail_acs_response(request, *args, **kwargs):
     but this can be simply changed so that PermissionDenied exception is raised instead.
     """
     failure_function = import_string(get_custom_setting('SAML_ACS_FAILURE_RESPONSE_FUNCTION',
-                                                        'djangosaml2_v0181post.acs_failures.template_failure'))
+                                                        'djangosaml2.acs_failures.template_failure'))
     return failure_function(request, *args, **kwargs)
